+++ noodles-cram/src/async/io/reader/header/container.rs
@@ -2,11 +2,12 @@ mod block;
 mod header;
 
 use bytes::BytesMut;
-use tokio::io::{self, AsyncRead, AsyncReadExt};
+use noodles_sam as sam;
+use tokio::io::{self, AsyncBufRead, AsyncBufReadExt, AsyncRead, AsyncReadExt, BufReader};
 
 use self::{block::read_block, header::read_header};
 
-pub async fn read_header_container<R>(reader: &mut R, buf: &mut BytesMut) -> io::Result<String>
+pub async fn read_raw_header_container<R>(reader: &mut R, buf: &mut BytesMut) -> io::Result<String>
 where
     R: AsyncRead + Unpin,
 {
@@ -37,12 +38,80 @@ where
     Ok(buf)
 }
 
+pub async fn read_header_container<R>(reader: &mut R, buf: &mut BytesMut) -> io::Result<sam::Header>
+where
+    R: AsyncRead + Unpin,
+{
+    let len = read_header(reader).await?;
+
+    buf.resize(len, 0);
+    reader.read_exact(buf).await?;
+
+    let buf = buf.split().freeze();
+    let mut reader = &buf[..];
+    read_sam_header(&mut reader).await
+}
+
+async fn read_sam_header<R>(reader: &mut R) -> io::Result<sam::Header>
+where
+    R: AsyncRead + Unpin,
+{
+    let mut block_reader = read_block(reader).await?;
+
+    let len = block_reader.read_i32_le().await.and_then(|n| {
+        u64::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    })?;
+
+    let mut parser = sam::header::Parser::default();
+
+    let mut header_reader = BufReader::new(block_reader.take(len));
+    let mut buf = Vec::new();
+
+    while read_line(&mut header_reader, &mut buf).await? != 0 {
+        parser
+            .parse_partial(&buf)
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+    }
+
+    Ok(parser.finish())
+}
+
+async fn read_line<R>(reader: &mut R, dst: &mut Vec<u8>) -> io::Result<usize>
+where
+    R: AsyncBufRead + Unpin,
+{
+    const LINE_FEED: u8 = b'\n';
+    const CARRIAGE_RETURN: u8 = b'\r';
+
+    dst.clear();
+
+    match reader.read_until(LINE_FEED, dst).await? {
+        0 => Ok(0),
+        n => {
+            if dst.ends_with(&[LINE_FEED]) {
+                dst.pop();
+
+                if dst.ends_with(&[CARRIAGE_RETURN]) {
+                    dst.pop();
+                }
+            }
+
+            Ok(n)
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[tokio::test]
     async fn test_read_raw_sam_header() -> io::Result<()> {
+        use sam::header::record::value::{
+            map::{self, header::Version},
+            Map,
+        };
+
         const RAW_HEADER: &str = "@HD\tVN:1.6\n";
 
         let mut src = vec![
@@ -57,9 +126,13 @@ mod tests {
         src.extend([0x00, 0x00, 0x00, 0x00]); // CRC32
 
         let mut reader = &src[..];
-        let actual = read_raw_sam_header(&mut reader).await?;
+        let actual = read_sam_header(&mut reader).await?;
 
-        assert_eq!(actual, RAW_HEADER);
+        let expected = sam::Header::builder()
+            .set_header(Map::<map::Header>::new(Version::new(1, 6)))
+            .build();
+
+        assert_eq!(actual, expected);
 
         Ok(())
     }
@@ -78,7 +151,7 @@ mod tests {
 
         let mut reader = &src[..];
         assert!(matches!(
-            read_raw_sam_header(&mut reader).await,
+            read_sam_header(&mut reader).await,
             Err(e) if e.kind() == io::ErrorKind::InvalidData
         ));
     }
@@ -93,7 +166,7 @@ mod tests {
         let mut reader = &src[..];
 
         assert!(matches!(
-            read_raw_sam_header(&mut reader).await,
+            read_sam_header(&mut reader).await,
             Err(e) if e.kind() == io::ErrorKind::InvalidData
         ));
     }